fix bug

+++ main.py
@@ -11,6 +11,15 @@ from format_srt import split_subtitles
 
 
 def add_summary_to_video(path, summary_name, text, video_name, output_video_name):
+    if not os.path.exists(path + video_name):
+        print("video not exists")
+        return
+    if not os.path.exists(path + summary_name):
+        print("summary not exists")
+        return
+    if os.path.exists(path + output_video_name):
+        print("output video exists")
+        return
     image_name = path + summary_name.split(".")[0] + ".png"
     create_text_image(text=text, image_name=image_name,
                       font_path='/System/Library/Fonts/STHeiti Light.ttc'
@@ -36,18 +45,24 @@ if __name__ == "__main__":
     translate_gpt(path, srt_name, input_language, output_language_srt, output_language, chunk_size)
 
     # 3. summarize subtitle
-    ans_text = summarize_with_langchain(path, srt_name)
     summary_filename = srt_name.split(".")[0] + "_summary.txt"
-    with open(path + summary_filename, "w") as f:
-        f.write(ans_text)
+    if os.path.exists(path + summary_filename):
+        print("summary file exists")
+    else:
+        ans_text = summarize_with_langchain(path, srt_name)
+        with open(path + summary_filename, "w") as f:
+            f.write(ans_text)
 
     # 4. translate summary
-    with open(path + summary_filename, "r") as f:
-        ans_text = f.read()
-    t_ans_text = translate_with_chatgpt(ans_text, "zh")
     summary_filename = output_language_srt.split(".")[0] + "_summary.txt"
-    with open(path + summary_filename, "w") as f:
-        f.write(t_ans_text)
+    if os.path.exists(path + summary_filename):
+        print("summary file exists")
+    else:
+        with open(path + summary_filename, "r") as f:
+            ans_text = f.read()
+        t_ans_text = translate_with_chatgpt(ans_text, "zh")
+        with open(path + summary_filename, "w") as f:
+            f.write(t_ans_text)
 
     # 5. format split subtitle
     # 5.1 split subtitle en
@@ -66,4 +81,6 @@ if __name__ == "__main__":
 
     # 6. add summary to video
     final_video_name = output_video_name.split(".")[0] + "_final.mp4"
+    with open(path + summary_filename, "r") as f:
+        t_ans_text = f.read()
     add_summary_to_video(path, summary_filename, t_ans_text, output_video_name, final_video_name)
